+++ wiki/wiki_md/about/changelog.md
@@ -29,6 +29,7 @@ This page documents the changes and fixes made in Quantum Things, compared to th
 - Fixed Runic Dust breaking all pieces at once in Creative Mode instead of one piece at a time.
 - Fixed Runic Dust dropping its item when broken in Creative Mode.
 - Fixed Precious Emeralds being removed from your inventory when giving them to villagers in creative mode.
+- Fixed beans not being able to be planted on non-vanilla blocks, even if the block could sustain plants.
 
 ## 1.0.2